Add URL-based routing and fix right navbar icons
Also do minor clean-up:
- organize files
- remove obsolete CSS rules
