+++ package.json
@@ -12,8 +12,10 @@
   "dependencies": {
     "@aws-amplify/ui-react": "^6.5.5",
     "aws-amplify": "^6.6.6",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^7.1.3"
   },
   "devDependencies": {
     "@aws-amplify/backend": "^1.5.0",
