+++ src/index.css
@@ -10,13 +10,8 @@ body {
 
 main {
   flex: 1 1 100%;
-  /*display: flex;*/
   flex-direction: column;
   align-items: stretch;
-}
-
-h4 {
-  text-align: left;
 }
 
 button {
@@ -31,9 +26,11 @@ button {
   transition: border-color 0.25s;
   color: white;
 }
+
 button:hover {
   border-color: #646cff;
 }
+
 button:focus,
 button:focus-visible {
   outline: 4px auto -webkit-focus-ring-color;
