+++ src/main.tsx
@@ -1,14 +1,18 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
-import "./index.css";
 import { Amplify } from "aws-amplify";
 import outputs from "../amplify_outputs.json";
+import { BrowserRouter } from 'react-router-dom';
+
+import "./index.css";
 
 Amplify.configure(outputs);
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <App />
+	<BrowserRouter>
+	  <App />
+	</BrowserRouter>
   </React.StrictMode>
 );
