+++ src/pages/Contact.tsx
@@ -1,0 +1,9 @@
+function Contact() {
+	return (
+	  <div>
+		<p>This is the Contact page</p>
+	  </div>
+	);
+};
+
+export default Contact;
