+++ src/pages/Help.tsx
@@ -1,0 +1,9 @@
+function Help() {
+	return (
+	  <div>
+		<p>This is the Help page</p>
+	  </div>
+	);
+};
+
+export default Help;
