+++ src/pages/Login.tsx
@@ -1,0 +1,9 @@
+function Login() {
+	return (
+	  <div>
+		<p>This is the Login page</p>
+	  </div>
+	);
+};
+
+export default Login;
